requirements.txt
google-cloud-speech==1.2.0
pydub==0.23.1